requirements.txt
duckduckgo-search>=6.0.0
pyyaml
typing_extensions
python-dateutil
requests
